Factoring out the two API client classes to separate files
