+++ Lombiq.OrchardCoreApiClient.Tester/Program.cs
@@ -1,3 +1,4 @@
+using Lombiq.OrchardCoreApiClient.Clients;
 using Lombiq.OrchardCoreApiClient.Models;
 using OrchardCore.ContentManagement;
 using System;
